+++ appointment_mediqe/accounts/models.py
@@ -8,30 +8,7 @@ from django.contrib.auth.models import (
 )
 import uuid
 
-
-# === Custom User Manager ===
-class CustomUserManager(BaseUserManager):
-    def create_user(self, email, password=None, **extra_fields):
-        """
-        Create and return a regular user with an email and password.
-        """
-        if not email:
-            raise ValueError("Email is required")
-        email = self.normalize_email(email)
-        user = self.model(email=email, **extra_fields)
-        user.set_password(password)  # Hash password
-        user.save(using=self._db)
-        return user
-
-    def create_superuser(self, email, password=None, **extra_fields):
-        """
-        Create and return a superuser with all permissions.
-        """
-        extra_fields.setdefault("is_staff", True)
-        extra_fields.setdefault("is_superuser", True)
-        return self.create_user(email, password, **extra_fields)
-
-
+from .managers import CustomUserManager
 # === Custom User Model ===
 class User(AbstractBaseUser, PermissionsMixin):
     ROLE_CHOICES = [
